+++ sweet/persist/Ucs2CharFilter.hpp
@@ -25,8 +25,8 @@ class SWEET_PERSIST_DECLSPEC Ucs2CharFilter
     public:
         void to_archive( const wchar_t* src_begin, const wchar_t* src_end, char* dest_begin, char* dest_end ) const;
         void to_memory( const char* src_begin, const char* src_end, wchar_t* dest_begin, wchar_t* dest_end ) const;
-        size_t to_archive_length( const wchar_t* begin, const wchar_t* end ) const;
-        size_t to_memory_length( const char* begin, const char* end ) const;
+        unsigned int to_archive_length( const wchar_t* begin, const wchar_t* end ) const;
+        unsigned int to_memory_length( const char* begin, const char* end ) const;
 };
 
 Ucs2CharFilter ucs2_char_filter();
